Rename StaticArgumentRename -> DrupalServiceRename
